GameExecutable: Continue refactoring. Rename some files.
