+++ Sources/SeriousSam/GUI/Menus/MenuActions.cpp
@@ -4,7 +4,7 @@
 #include <Engine/Build.h>
 
 #include "MenuManager.h"
-#include "Menu_starters.h"
+#include "MenuStarters.h"
 #include "MenuStuff.h"
 #include "GUI/Components/MenuGadget.h"
 #include "LevelInfo.h"
