+++ Sources/SeriousSam/GUI/Menus/MenuActions.h
@@ -1,0 +1,31 @@
+/* Copyright (c) 2002-2012 Croteam Ltd. All rights reserved. */
+
+#ifndef SE_INCL_MENU_ACTIONS_H
+#define SE_INCL_MENU_ACTIONS_H
+#ifdef PRAGMA_ONCE
+  #pragma once
+#endif
+
+
+void InitActionsForAudioOptionsMenu();
+void InitActionsForConfirmMenu();
+void InitActionsForControlsMenu();
+void InitActionsForCustomizeAxisMenu();
+void InitActionsForMainMenu();
+void InitActionsForInGameMenu();
+void InitActionsForNetworkMenu();
+void InitActionsForNetworkJoinMenu();
+void InitActionsForNetworkOpenMenu();
+void InitActionsForNetworkStartMenu();
+void InitActionsForOptionsMenu();
+void InitActionsForPlayerProfileMenu();
+void InitActionsForSelectPlayersMenu();
+void InitActionsForServersMenu();
+void InitActionsForSinglePlayerMenu();
+void InitActionsForSinglePlayerNewMenu();
+void InitActionsForSplitScreenMenu();
+void InitActionsForSplitStartMenu();
+void InitActionsForVideoOptionsMenu();
+void InitActionsForVarMenu();
+
+#endif  /* include-once check. */
