+++ Sources/SeriousSam/GUI/Menus/MenuStarters.h
