+++ Sources/SeriousSam/SeriousSam.vcxproj
@@ -202,10 +202,11 @@ copy Release\$(TargetName).map $(SolutionDir)..\Bin\ &gt;nul </Command>
     <ClCompile Include="GUI\Menus\MCustomizeKeyboard.cpp" />
     <ClCompile Include="GUI\Menus\MDisabled.cpp" />
     <ClCompile Include="GUI\Menus\Menu.cpp" />
+    <ClCompile Include="GUI\Menus\MenuStartersAF.cpp" />
     <ClCompile Include="GUI\Menus\MenuManager.cpp" />
     <ClCompile Include="GUI\Menus\MenuPrinting.cpp" />
     <ClCompile Include="GUI\Menus\MenuStuff.cpp" />
-    <ClCompile Include="GUI\Menus\Menu_actions.cpp" />
+    <ClCompile Include="GUI\Menus\MenuActions.cpp" />
     <ClCompile Include="GUI\Menus\MHighScore.cpp" />
     <ClCompile Include="GUI\Menus\MInGame.cpp" />
     <ClCompile Include="GUI\Menus\MLevels.cpp" />
@@ -252,10 +253,12 @@ copy Release\$(TargetName).map $(SolutionDir)..\Bin\ &gt;nul </Command>
     <ClInclude Include="GUI\Menus\MCustomizeKeyboard.h" />
     <ClInclude Include="GUI\Menus\MDisabled.h" />
     <ClInclude Include="GUI\Menus\Menu.h" />
+    <ClInclude Include="GUI\Menus\MenuActions.h" />
     <ClInclude Include="GUI\Menus\MenuManager.h" />
     <ClInclude Include="GUI\Menus\MenuPrinting.h" />
+    <ClInclude Include="GUI\Menus\MenuStartersAF.h" />
     <ClInclude Include="GUI\Menus\MenuStuff.h" />
-    <ClInclude Include="GUI\Menus\Menu_starters.h" />
+    <ClInclude Include="GUI\Menus\MenuStarters.h" />
     <ClInclude Include="GUI\Menus\MHighScore.h" />
     <ClInclude Include="GUI\Menus\MInGame.h" />
     <ClInclude Include="GUI\Menus\MLevels.h" />
@@ -276,7 +279,7 @@ copy Release\$(TargetName).map $(SolutionDir)..\Bin\ &gt;nul </Command>
     <ClInclude Include="GUI\Menus\MSplitStart.h" />
     <ClInclude Include="GUI\Menus\MVar.h" />
     <ClInclude Include="GUI\Menus\MVideoOptions.h" />
-    <ClCompile Include="GUI\Menus\Menu_starters.cpp" />
+    <ClCompile Include="GUI\Menus\MenuStarters.cpp" />
     <ClCompile Include="SeriousSam.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Use</PrecompiledHeader>
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">StdH.h</PrecompiledHeaderFile>
